Take every clippy warning in the tree to zero
Eleven warnings across five crates, none of them alarming on their own,
but a reviewer running clippy should see a clean tree rather than have to
learn which eleven are expected. Most are mechanical: two no-op casts in
the SAPI shim, a let-and-return in the hook, an unwrap after is_none, and
a missing safety doc.

Two changed behaviour, which is exactly what should not hide inside a
cleanup, so they are called out here.

benchmark.rs guarded its speed calculation with !(elapsed > 0.0), written
negated on purpose to reject NaN - the naive "fix" clippy suggests would
have been a regression. But that form lets +inf through and divides to the
absurd speed the comment says it is preventing. It now requires finite and
positive, which is clippy-clean and strictly stronger. Both values are read
off the pipe, so neither is trusted.

The panel's verify progress computed pct and frac from `total` on adjacent
lines and disagreed about zero: pct said 100 while frac said 0.0, so a
manifest of zero-byte files showed a full percentage over an empty bar.
Decided once now. Codex pointed out the first version of the comment named
an unreachable case - an EMPTY manifest never reaches the callback at all,
because download::verify only calls it from inside its per-file loop - so
the comment describes the case that actually occurs.

DllGetClassObject gained the safety doc clippy asked for, and then had to
gain an E_POINTER check to make that doc true: the documented contract said
a null ppv was tolerated, but the only thing the code could do with one was
hand it to query, whose own contract requires somewhere to write. The null
handling predates this change; the false promise did not. Rejecting outright
matches what the engine's other out-pointer entry points already do.

No cargo fmt. The repo is not rustfmt-clean and there is no rustfmt.toml,
so running it rewrites ~970 lines across nine files and the line endings
with them, which would bury every change above.

+++ kokoro-sapi/src/engine.rs
@@ -284,7 +284,7 @@ impl ISpTTSEngine_Impl for KokoroEngine_Impl {
             return E_OUTOFMEMORY;
         }
         *wfx = WAVEFORMATEX {
-            wFormatTag: WAVE_FORMAT_PCM as u16,
+            wFormatTag: WAVE_FORMAT_PCM,
             nChannels: 1,
             nSamplesPerSec: SAMPLE_RATE,
             wBitsPerSample: 16,
